renamed modules to lower-case for clearer import structures
