+++ src/imageprovider/ImageProviderConfig.py
