+++ src/slicer/SlicerConfig.py
